rearrange tsdf headers, remove old tsdf stuff, start altering tsdf python port and SDF2SDF
