+++ src/tsdf/ewa_3d_viz.cpp
@@ -29,7 +29,7 @@
 //local
 #include "common.hpp"
 #include <lsf_config.h>
-#include "ewa.hpp"
+#include "ewa_viz.hpp"
 #include "ewa_common.hpp"
 #include "../math/conics.hpp"
