+++ tests/data/test_data_tsdf.hpp
@@ -96,22 +96,22 @@ static eig::MatrixXf out_sdf_chunk =
 static eig::Tensor<float, 3> TSDF_slice01 =
 		[] {
 			float data[] = {
-					-0.20475684f, -0.13041718f, -0.05159303f, 0.01631528f, 0.09734481f, 0.18223374f, 0.25901046f, 0.3345288f, 0.4215054f, 0.5017995f, 0.57773066f, 0.67372465f, 0.7540442f, 0.8337743f, 0.9401298f, 1.0f,
-					-0.2718329f, -0.19898264f, -0.11795311f, -0.05366355f, 0.03163889f, 0.11868029f, 0.18068849f, 0.2674229f, 0.35445613f, 0.42136607f, 0.51035875f, 0.6072543f, 0.6730221f, 0.76607686f, 0.8696056f, 0.9367048f,
-					-0.33940297f, -0.26726198f, -0.19352509f, -0.11860802f, -0.0351347f, 0.04477277f, 0.11306702f, 0.19761248f, 0.28388125f, 0.34753007f, 0.44209582f, 0.5338877f, 0.59785986f, 0.69736683f, 0.7920019f, 0.86319065f,
-					-0.41053814f, -0.33195686f, -0.26949567f, -0.18563195f, -0.10349526f, -0.03297701f, 0.04537329f, 0.12810154f, 0.20457803f, 0.28029305f, 0.3743246f, 0.4515655f, 0.529439f, 0.6300881f, 0.7094442f, 0.7964156f,
-					-0.47973615f, -0.40533316f, -0.3329001f, -0.25372428f, -0.17178132f, -0.10733827f, -0.0254929f, 0.06117373f, 0.12537389f, 0.21193548f, 0.30558106f, 0.37155968f, 0.46049577f, 0.56024337f, 0.6278216f, 0.72762364f,
-					-0.54754835f, -0.48085f, -0.40127334f, -0.32249835f, -0.24583338f, -0.17565785f, -0.09573101f, -0.01198053f, 0.05252584f, 0.14263166f, 0.23218466f, 0.29368994f, 0.3898285f, 0.484772f, 0.5527272f, 0.6566673f,
-					-0.6144799f, -0.54812276f, -0.4697546f, -0.3909312f, -0.32430065f, -0.2446696f, -0.16611888f, -0.09179189f, -0.01491904f, 0.07018074f, 0.15120803f, 0.22628976f, 0.3201082f, 0.40280667f, 0.48454177f, 0.58679205f,
-					-0.69139594f, -0.61513406f, -0.5401842f, -0.4608303f, -0.3944121f, -0.31756383f, -0.23505314f, -0.17100646f, -0.08574425f, 0.00108704f, 0.06889998f, 0.1552522f, 0.24996398f, 0.31985265f, 0.41284856f, 0.5157612f,
-					-0.76291186f, -0.6857901f, -0.6103374f, -0.5386583f, -0.46513748f, -0.38871688f, -0.30887496f, -0.2414055f, -0.15761329f, -0.06929486f, -0.00930652f, 0.08366852f, 0.17710625f, 0.24074091f, 0.34142134f, 0.43653396f,
-					-0.82926893f, -0.7568284f, -0.6821744f, -0.6148077f, -0.5358427f, -0.45973283f, -0.38855296f, -0.3113508f, -0.22937803f, -0.15131308f, -0.07774457f, 0.00998154f, 0.09599998f, 0.17126425f, 0.26991662f, 0.35248694f,
-					-0.8988074f, -0.8265122f, -0.7519416f, -0.6832592f, -0.60939413f, -0.5306705f, -0.46718267f, -0.38460863f, -0.30043197f, -0.23241787f, -0.14970823f, -0.06056577f, 0.01350045f, 0.09886324f, 0.19591151f, 0.26573387f,
-					-0.97098565f, -0.89905554f, -0.8297361f, -0.7546924f, -0.68361086f, -0.6053954f, -0.53770316f, -0.45659837f, -0.37168336f, -0.31055582f, -0.22116004f, -0.1305498f, -0.06811767f, 0.02796873f, 0.12376829f, 0.18438174f,
-					-1.0f, -0.96868193f, -0.9008161f, -0.82786226f, -0.7554948f, -0.68548465f, -0.607904f, -0.5285464f, -0.45156848f, -0.38096604f, -0.2947412f, -0.20934565f, -0.14037861f, -0.04654675f, 0.04146173f, 0.10772794f,
-					-1.0f, -1.0f, -0.97198033f, -0.9039304f, -0.8274629f, -0.7603347f, -0.6806403f, -0.60253733f, -0.5349218f, -0.45473945f, -0.37067902f, -0.29290763f, -0.21186395f, -0.12094676f, -0.04442408f, 0.03523007f,
-					-1.0f, -1.0f, -1.0f, -0.9779266f, -0.9029112f, -0.832349f, -0.75637394f, -0.6757326f, -0.6126381f, -0.52764934f, -0.44117716f, -0.37548465f, -0.28660965f, -0.19398032f, -0.12813507f, -0.03899559f,
-					-1.0f, -1.0f, -1.0f, -1.0f, -0.98226213f, -0.90591604f, -0.83188707f, -0.75543445f, -0.68388426f, -0.60274744f, -0.51609796f, -0.45198423f, -0.3620803f, -0.27120334f, -0.20937918f, -0.11253878f
+				-0.20475684f, -0.13041718f, -0.05159303f, 0.01631528f, 0.09734481f, 0.18223374f, 0.25901046f, 0.3345288f, 0.4215054f, 0.5017995f, 0.57773066f, 0.67372465f, 0.7540442f, 0.8337743f, 0.9401298f, 1.0f,
+				-0.2718329f, -0.19898264f, -0.11795311f, -0.05366355f, 0.03163889f, 0.11868029f, 0.18068849f, 0.2674229f, 0.35445613f, 0.42136607f, 0.51035875f, 0.6072543f, 0.6730221f, 0.76607686f, 0.8696056f, 0.9367048f,
+				-0.33940297f, -0.26726198f, -0.19352509f, -0.11860802f, -0.0351347f, 0.04477277f, 0.11306702f, 0.19761248f, 0.28388125f, 0.34753007f, 0.44209582f, 0.5338877f, 0.59785986f, 0.69736683f, 0.7920019f, 0.86319065f,
+				-0.41053814f, -0.33195686f, -0.26949567f, -0.18563195f, -0.10349526f, -0.03297701f, 0.04537329f, 0.12810154f, 0.20457803f, 0.28029305f, 0.3743246f, 0.4515655f, 0.529439f, 0.6300881f, 0.7094442f, 0.7964156f,
+				-0.47973615f, -0.40533316f, -0.3329001f, -0.25372428f, -0.17178132f, -0.10733827f, -0.0254929f, 0.06117373f, 0.12537389f, 0.21193548f, 0.30558106f, 0.37155968f, 0.46049577f, 0.56024337f, 0.6278216f, 0.72762364f,
+				-0.54754835f, -0.48085f, -0.40127334f, -0.32249835f, -0.24583338f, -0.17565785f, -0.09573101f, -0.01198053f, 0.05252584f, 0.14263166f, 0.23218466f, 0.29368994f, 0.3898285f, 0.484772f, 0.5527272f, 0.6566673f,
+				-0.6144799f, -0.54812276f, -0.4697546f, -0.3909312f, -0.32430065f, -0.2446696f, -0.16611888f, -0.09179189f, -0.01491904f, 0.07018074f, 0.15120803f, 0.22628976f, 0.3201082f, 0.40280667f, 0.48454177f, 0.58679205f,
+				-0.69139594f, -0.61513406f, -0.5401842f, -0.4608303f, -0.3944121f, -0.31756383f, -0.23505314f, -0.17100646f, -0.08574425f, 0.00108704f, 0.06889998f, 0.1552522f, 0.24996398f, 0.31985265f, 0.41284856f, 0.5157612f,
+				-0.76291186f, -0.6857901f, -0.6103374f, -0.5386583f, -0.46513748f, -0.38871688f, -0.30887496f, -0.2414055f, -0.15761329f, -0.06929486f, -0.00930652f, 0.08366852f, 0.17710625f, 0.24074091f, 0.34142134f, 0.43653396f,
+				-0.82926893f, -0.7568284f, -0.6821744f, -0.6148077f, -0.5358427f, -0.45973283f, -0.38855296f, -0.3113508f, -0.22937803f, -0.15131308f, -0.07774457f, 0.00998154f, 0.09599998f, 0.17126425f, 0.26991662f, 0.35248694f,
+				-0.8988074f, -0.8265122f, -0.7519416f, -0.6832592f, -0.60939413f, -0.5306705f, -0.46718267f, -0.38460863f, -0.30043197f, -0.23241787f, -0.14970823f, -0.06056577f, 0.01350045f, 0.09886324f, 0.19591151f, 0.26573387f,
+				-0.97098565f, -0.89905554f, -0.8297361f, -0.7546924f, -0.68361086f, -0.6053954f, -0.53770316f, -0.45659837f, -0.37168336f, -0.31055582f, -0.22116004f, -0.1305498f, -0.06811767f, 0.02796873f, 0.12376829f, 0.18438174f,
+				-1.0f, -0.96868193f, -0.9008161f, -0.82786226f, -0.7554948f, -0.68548465f, -0.607904f, -0.5285464f, -0.45156848f, -0.38096604f, -0.2947412f, -0.20934565f, -0.14037861f, -0.04654675f, 0.04146173f, 0.10772794f,
+				-1.0f, -1.0f, -0.97198033f, -0.9039304f, -0.8274629f, -0.7603347f, -0.6806403f, -0.60253733f, -0.5349218f, -0.45473945f, -0.37067902f, -0.29290763f, -0.21186395f, -0.12094676f, -0.04442408f, 0.03523007f,
+				-1.0f, -1.0f, -1.0f, -0.9779266f, -0.9029112f, -0.832349f, -0.75637394f, -0.6757326f, -0.6126381f, -0.52764934f, -0.44117716f, -0.37548465f, -0.28660965f, -0.19398032f, -0.12813507f, -0.03899559f,
+				-1.0f, -1.0f, -1.0f, -1.0f, -0.98226213f, -0.90591604f, -0.83188707f, -0.75543445f, -0.68388426f, -0.60274744f, -0.51609796f, -0.45198423f, -0.3620803f, -0.27120334f, -0.20937918f, -0.11253878f
 			};
 
 			auto mapped_t = Eigen::TensorMap<Eigen::Tensor<float, 3, Eigen::RowMajor>>(data, 16, 1, 16);
@@ -122,22 +122,22 @@ static eig::Tensor<float, 3> TSDF_slice01 =
 static eig::Tensor<float, 3> TSDF_slice02 =
 		[] {
 			float data[] = {
-					-0.15828474f, -0.11724376f, -0.10635137f, 0.0347948f, 0.13013457f, 0.13970095f, 0.2607098f, 0.37702486f, 0.38655585f, 0.4914882f, 0.6236901f, 0.63738084f, 0.7511944f, 0.8743434f, 0.89005965f, 1.0f,
-					-0.2258139f, -0.2199551f, -0.15794252f, -0.00265987f, 0.03356163f, 0.06181556f, 0.22034676f, 0.28386486f, 0.29687205f, 0.4515414f, 0.533245f, 0.54962367f, 0.7096594f, 0.7776409f, 0.8098362f, 0.9918554f,
-					-0.31726792f, -0.31567535f, -0.19318739f, -0.07427778f, -0.06489212f, 0.01729878f, 0.16743435f, 0.18366536f, 0.23672199f, 0.40936098f, 0.43442923f, 0.48488897f, 0.65332645f, 0.67893976f, 0.75372607f, 0.92021626f,
-					-0.41854447f, -0.3823676f, -0.23076715f, -0.16607293f, -0.15566197f, -0.01788376f, 0.082306f, 0.08623048f, 0.19767742f, 0.32805297f, 0.33478448f, 0.44285813f, 0.5730631f, 0.5839522f, 0.7176227f, 0.830434f,
-					-0.51883006f, -0.42022684f, -0.28082898f, -0.26674107f, -0.21839245f, -0.05833106f, -0.01604484f, 0.00105352f, 0.16258742f, 0.23426425f, 0.24564722f, 0.40389f, 0.47854367f, 0.49812043f, 0.6721107f, 0.73466957f,
-					-0.6033763f, -0.4579158f, -0.3659076f, -0.36507088f, -0.2567028f, -0.12278296f, -0.11890986f, -0.05184137f, 0.11196412f, 0.13258268f, 0.17926085f, 0.3554237f, 0.37924826f, 0.4338902f, 0.6157429f, 0.63608646f,
-					-0.645462f, -0.49682447f, -0.46659848f, -0.44712663f, -0.29905558f, -0.21631671f, -0.21445934f, -0.0931825f, 0.02886167f, 0.0339981f, 0.13835897f, 0.27350315f, 0.28232044f, 0.3953811f, 0.5281225f, 0.54116374f,
-					-0.6869416f, -0.5688098f, -0.5686373f, -0.49458835f, -0.3414111f, -0.31428707f, -0.29624635f, -0.13149002f, -0.063048f, -0.05861344f, 0.09790006f, 0.17983738f, 0.18881369f, 0.35134864f, 0.43510342f, 0.45342973f,
-					-0.72327024f, -0.66508883f, -0.661359f, -0.52946675f, -0.41768035f, -0.41714862f, -0.34437734f, -0.18377991f, -0.16648252f, -0.12047808f, 0.05118029f, 0.07906744f, 0.11693326f, 0.30343905f, 0.3356216f, 0.38579437f,
-					-0.7763358f, -0.7674736f, -0.7283486f, -0.57718f, -0.51306856f, -0.514597f, -0.38262054f, -0.26476237f, -0.26809445f, -0.16572699f, -0.02654833f, -0.02175728f, 0.07301583f, 0.22863467f, 0.23383684f, 0.3388197f,
-					-0.8644161f, -0.8672254f, -0.7708329f, -0.6276073f, -0.6174873f, -0.5900806f, -0.4260473f, -0.36359546f, -0.3600693f, -0.20942566f, -0.12119135f, -0.1146424f, 0.0320527f, 0.13398415f, 0.14114645f, 0.2985325f,
-					-0.96556437f, -0.9566214f, -0.8115571f, -0.71163124f, -0.7212831f, -0.6320864f, -0.47746128f, -0.46737003f, -0.42747727f, -0.2539534f, -0.21941717f, -0.19709189f, -0.01466543f, 0.03628113f, 0.05920541f, 0.23854096f,
-					-1.0f, -1.0f, -0.85546386f, -0.81321603f, -0.81221074f, -0.6737332f, -0.5629941f, -0.57266414f, -0.47133508f, -0.32217124f, -0.32374653f, -0.2472215f, -0.07647318f, -0.06676186f, 0.00130505f, 0.16856915f,
-					-1.0f, -1.0f, -0.92087454f, -0.9182008f, -0.87941104f, -0.7177957f, -0.661619f, -0.6644632f, -0.51623654f, -0.41769585f, -0.42159325f, -0.28783286f, -0.16332234f, -0.167712f, -0.03794361f, 0.0806923f,
-					-1.0f, -1.0f, -1.0f, -1.0f, -0.9208766f, -0.7725691f, -0.7679372f, -0.7299541f, -0.5587841f, -0.519628f, -0.5087995f, -0.3345998f, -0.2650984f, -0.25806686f, -0.09272385f, -0.019745f,
-					-1.0f, -1.0f, -1.0f, -1.0f, -0.9690759f, -0.8593376f, -0.8730626f, -0.77391493f, -0.6238807f, -0.6228114f, -0.56681186f, -0.38827598f, -0.36741194f, -0.32917243f, -0.14534502f, -0.12005218f
+				-0.15828474f, -0.11724376f, -0.10635137f, 0.0347948f, 0.13013457f, 0.13970095f, 0.2607098f, 0.37702486f, 0.38655585f, 0.4914882f, 0.6236901f, 0.63738084f, 0.7511944f, 0.8743434f, 0.89005965f, 1.0f,
+				-0.2258139f, -0.2199551f, -0.15794252f, -0.00265987f, 0.03356163f, 0.06181556f, 0.22034676f, 0.28386486f, 0.29687205f, 0.4515414f, 0.533245f, 0.54962367f, 0.7096594f, 0.7776409f, 0.8098362f, 0.9918554f,
+				-0.31726792f, -0.31567535f, -0.19318739f, -0.07427778f, -0.06489212f, 0.01729878f, 0.16743435f, 0.18366536f, 0.23672199f, 0.40936098f, 0.43442923f, 0.48488897f, 0.65332645f, 0.67893976f, 0.75372607f, 0.92021626f,
+				-0.41854447f, -0.3823676f, -0.23076715f, -0.16607293f, -0.15566197f, -0.01788376f, 0.082306f, 0.08623048f, 0.19767742f, 0.32805297f, 0.33478448f, 0.44285813f, 0.5730631f, 0.5839522f, 0.7176227f, 0.830434f,
+				-0.51883006f, -0.42022684f, -0.28082898f, -0.26674107f, -0.21839245f, -0.05833106f, -0.01604484f, 0.00105352f, 0.16258742f, 0.23426425f, 0.24564722f, 0.40389f, 0.47854367f, 0.49812043f, 0.6721107f, 0.73466957f,
+				-0.6033763f, -0.4579158f, -0.3659076f, -0.36507088f, -0.2567028f, -0.12278296f, -0.11890986f, -0.05184137f, 0.11196412f, 0.13258268f, 0.17926085f, 0.3554237f, 0.37924826f, 0.4338902f, 0.6157429f, 0.63608646f,
+				-0.645462f, -0.49682447f, -0.46659848f, -0.44712663f, -0.29905558f, -0.21631671f, -0.21445934f, -0.0931825f, 0.02886167f, 0.0339981f, 0.13835897f, 0.27350315f, 0.28232044f, 0.3953811f, 0.5281225f, 0.54116374f,
+				-0.6869416f, -0.5688098f, -0.5686373f, -0.49458835f, -0.3414111f, -0.31428707f, -0.29624635f, -0.13149002f, -0.063048f, -0.05861344f, 0.09790006f, 0.17983738f, 0.18881369f, 0.35134864f, 0.43510342f, 0.45342973f,
+				-0.72327024f, -0.66508883f, -0.661359f, -0.52946675f, -0.41768035f, -0.41714862f, -0.34437734f, -0.18377991f, -0.16648252f, -0.12047808f, 0.05118029f, 0.07906744f, 0.11693326f, 0.30343905f, 0.3356216f, 0.38579437f,
+				-0.7763358f, -0.7674736f, -0.7283486f, -0.57718f, -0.51306856f, -0.514597f, -0.38262054f, -0.26476237f, -0.26809445f, -0.16572699f, -0.02654833f, -0.02175728f, 0.07301583f, 0.22863467f, 0.23383684f, 0.3388197f,
+				-0.8644161f, -0.8672254f, -0.7708329f, -0.6276073f, -0.6174873f, -0.5900806f, -0.4260473f, -0.36359546f, -0.3600693f, -0.20942566f, -0.12119135f, -0.1146424f, 0.0320527f, 0.13398415f, 0.14114645f, 0.2985325f,
+				-0.96556437f, -0.9566214f, -0.8115571f, -0.71163124f, -0.7212831f, -0.6320864f, -0.47746128f, -0.46737003f, -0.42747727f, -0.2539534f, -0.21941717f, -0.19709189f, -0.01466543f, 0.03628113f, 0.05920541f, 0.23854096f,
+				-1.0f, -1.0f, -0.85546386f, -0.81321603f, -0.81221074f, -0.6737332f, -0.5629941f, -0.57266414f, -0.47133508f, -0.32217124f, -0.32374653f, -0.2472215f, -0.07647318f, -0.06676186f, 0.00130505f, 0.16856915f,
+				-1.0f, -1.0f, -0.92087454f, -0.9182008f, -0.87941104f, -0.7177957f, -0.661619f, -0.6644632f, -0.51623654f, -0.41769585f, -0.42159325f, -0.28783286f, -0.16332234f, -0.167712f, -0.03794361f, 0.0806923f,
+				-1.0f, -1.0f, -1.0f, -1.0f, -0.9208766f, -0.7725691f, -0.7679372f, -0.7299541f, -0.5587841f, -0.519628f, -0.5087995f, -0.3345998f, -0.2650984f, -0.25806686f, -0.09272385f, -0.019745f,
+				-1.0f, -1.0f, -1.0f, -1.0f, -0.9690759f, -0.8593376f, -0.8730626f, -0.77391493f, -0.6238807f, -0.6228114f, -0.56681186f, -0.38827598f, -0.36741194f, -0.32917243f, -0.14534502f, -0.12005218f
 			};
 
 			auto mapped_t = Eigen::TensorMap<Eigen::Tensor<float, 3, Eigen::RowMajor>>(data, 16, 1, 16);
@@ -149,22 +149,22 @@ static eig::MatrixXf expected_tsdf_field01 =
 		[] {
 			eig::MatrixXf expected_tsdf_field01(16, 16);
 			expected_tsdf_field01 <<
-					-0.0249997f, 0.0499994f, 0.299999f, 0.2f, 0.6f, 0.399999f, 0.7f, 0.7f, 0.9f, 0.974999f, 1.f , 1.f , 1.f , 1.f , 1.f , 1.f ,
-					-0.125f, -0.0500008f, 0.199999f, 0.1f, 0.499999f, 0.25f, 0.6f, 0.6f, 0.8f, 0.874999f, 1.f , 1.f , 1.f , 1.f , 1.f , 1.f ,
-					-0.275001f, -0.150001f, 0.0999987f, 0.f , 0.399999f, 0.15f, 0.499999f, 0.499999f, 0.7f, 0.774999f, 0.925f, 1.f , 1.f , 1.f , 1.f , 1.f ,
-					-0.375f, -0.174999f, 0.f , -0.0999987f, 0.300001f, 0.0500008f, 0.400001f, 0.400001f, 0.600001f, 0.675f, 0.825001f, 0.975001f, 0.975001f, 1.f , 1.f , 1.f ,
-					-0.475f, -0.274999f, -0.15f, -0.199999f, 0.2f, -0.0499994f, 0.400001f, 0.300001f, 0.500001f, 0.575f, 0.625001f, 0.875001f, 0.875001f, 1.f , 1.f , 1.f ,
-					-0.575f, -0.375f, -0.25f, -0.225f, 0.1f, -0.15f, 0.300001f, 0.2f, 0.400001f, 0.475f, 0.525001f, 0.775f, 0.775f, 1.f , 0.95f, 1.f ,
-					-0.725f, -0.475f, -0.35f, -0.325f, 0.f , -0.25f, 0.2f, 0.1f, 0.300001f, 0.375f, 0.425f, 0.675f, 0.675f, 0.95f, 0.849999f, 1.f ,
-					-0.825f, -0.499999f, -0.45f, -0.425f, -0.1f, -0.35f, 0.1f, 0.f , 0.2f, 0.274999f, 0.325f, 0.575f, 0.575f, 0.849999f, 0.749999f, 1.f ,
-					-0.925f, -0.6f, -0.55f, -0.525001f, -0.25f, -0.45f, 0.f , -0.1f, 0.1f, 0.174999f, 0.225f, 0.475f, 0.375f, 0.749999f, 0.649999f, 1.f ,
-					-1.f, -0.7f, -0.7f, -0.625001f, -0.35f, -0.55f, -0.1f, -0.2f, 0.f , 0.074999f, 0.125f, 0.375f, 0.274999f, 0.649999f, 0.475f, 0.925f,
-					-1.f, -0.8f, -0.8f, -0.725001f, -0.45f, -0.650001f, -0.2f, -0.300001f, -0.1f, -0.0250012f, 0.0249997f, 0.274999f, 0.174999f, 0.549999f, 0.375f, 0.825f,
-					-1.f, -0.899999f, -0.899999f, -0.825f, -0.549999f, -0.749999f, -0.299999f, -0.399999f, -0.199999f, -0.125f, -0.074999f, 0.175001f, 0.0750005f, 0.45f, 0.275001f, 0.725001f,
-					-1.f, -0.9f, -0.999999f, -0.849999f, -0.649999f, -0.749999f, -0.399999f, -0.549999f, -0.299999f, -0.199999f, -0.174999f, 0.125f, -0.0249997f, 0.400001f, 0.175001f, 0.625001f,
-					-1.f, -1.f , -1.f , -0.95f, -0.749999f, -0.849999f, -0.499999f, -0.649999f, -0.399999f, -0.299999f, -0.274999f, 0.0249997f, -0.125f, 0.300001f, 0.0750005f, 0.525001f,
-					-1.f, -1.f , -1.f , -1.f , -0.849999f, -0.95f, -0.6f, -0.749999f, -0.499999f, -0.399999f, -0.375f, -0.0750005f, -0.225f, 0.2f, -0.0249997f, 0.425f,
-					-1.f, -1.f , -1.f , -1.f , -0.95f, -1.f , -0.7f, -0.849999f, -0.6f, -0.499999f, -0.475f, -0.175001f, -0.325f, 0.1f, -0.1f, 0.325f;
+			-0.0249997f, 0.0499994f, 0.299999f, 0.2f, 0.6f, 0.399999f, 0.7f, 0.7f, 0.9f, 0.974999f, 1.f , 1.f , 1.f , 1.f , 1.f , 1.f ,
+			-0.125f, -0.0500008f, 0.199999f, 0.1f, 0.499999f, 0.25f, 0.6f, 0.6f, 0.8f, 0.874999f, 1.f , 1.f , 1.f , 1.f , 1.f , 1.f ,
+			-0.275001f, -0.150001f, 0.0999987f, 0.f , 0.399999f, 0.15f, 0.499999f, 0.499999f, 0.7f, 0.774999f, 0.925f, 1.f , 1.f , 1.f , 1.f , 1.f ,
+			-0.375f, -0.174999f, 0.f , -0.0999987f, 0.300001f, 0.0500008f, 0.400001f, 0.400001f, 0.600001f, 0.675f, 0.825001f, 0.975001f, 0.975001f, 1.f , 1.f , 1.f ,
+			-0.475f, -0.274999f, -0.15f, -0.199999f, 0.2f, -0.0499994f, 0.400001f, 0.300001f, 0.500001f, 0.575f, 0.625001f, 0.875001f, 0.875001f, 1.f , 1.f , 1.f ,
+			-0.575f, -0.375f, -0.25f, -0.225f, 0.1f, -0.15f, 0.300001f, 0.2f, 0.400001f, 0.475f, 0.525001f, 0.775f, 0.775f, 1.f , 0.95f, 1.f ,
+			-0.725f, -0.475f, -0.35f, -0.325f, 0.f , -0.25f, 0.2f, 0.1f, 0.300001f, 0.375f, 0.425f, 0.675f, 0.675f, 0.95f, 0.849999f, 1.f ,
+			-0.825f, -0.499999f, -0.45f, -0.425f, -0.1f, -0.35f, 0.1f, 0.f , 0.2f, 0.274999f, 0.325f, 0.575f, 0.575f, 0.849999f, 0.749999f, 1.f ,
+			-0.925f, -0.6f, -0.55f, -0.525001f, -0.25f, -0.45f, 0.f , -0.1f, 0.1f, 0.174999f, 0.225f, 0.475f, 0.375f, 0.749999f, 0.649999f, 1.f ,
+			-1.f, -0.7f, -0.7f, -0.625001f, -0.35f, -0.55f, -0.1f, -0.2f, 0.f , 0.074999f, 0.125f, 0.375f, 0.274999f, 0.649999f, 0.475f, 0.925f,
+			-1.f, -0.8f, -0.8f, -0.725001f, -0.45f, -0.650001f, -0.2f, -0.300001f, -0.1f, -0.0250012f, 0.0249997f, 0.274999f, 0.174999f, 0.549999f, 0.375f, 0.825f,
+			-1.f, -0.899999f, -0.899999f, -0.825f, -0.549999f, -0.749999f, -0.299999f, -0.399999f, -0.199999f, -0.125f, -0.074999f, 0.175001f, 0.0750005f, 0.45f, 0.275001f, 0.725001f,
+			-1.f, -0.9f, -0.999999f, -0.849999f, -0.649999f, -0.749999f, -0.399999f, -0.549999f, -0.299999f, -0.199999f, -0.174999f, 0.125f, -0.0249997f, 0.400001f, 0.175001f, 0.625001f,
+			-1.f, -1.f , -1.f , -0.95f, -0.749999f, -0.849999f, -0.499999f, -0.649999f, -0.399999f, -0.299999f, -0.274999f, 0.0249997f, -0.125f, 0.300001f, 0.0750005f, 0.525001f,
+			-1.f, -1.f , -1.f , -1.f , -0.849999f, -0.95f, -0.6f, -0.749999f, -0.499999f, -0.399999f, -0.375f, -0.0750005f, -0.225f, 0.2f, -0.0249997f, 0.425f,
+			-1.f, -1.f , -1.f , -1.f , -0.95f, -1.f , -0.7f, -0.849999f, -0.6f, -0.499999f, -0.475f, -0.175001f, -0.325f, 0.1f, -0.1f, 0.325f;
 			return expected_tsdf_field01;
 		}();
 
@@ -172,22 +172,69 @@ static eig::MatrixXf expected_tsdf_field02 =
 		[] {
 			eig::MatrixXf expected_tsdf_field01(16, 16);
 			expected_tsdf_field01 <<
-					-0.125f, -0.0500008f, 0.199999f, 0.1f, 0.499999f, 0.25f, 0.6f, 0.6f, 0.8f, 0.874999f, 1.f, 1.f, 1.f, 1.f, 1.f, 1.f,
-					-0.275001f, -0.150001f, 0.0999987f, 0.f, 0.399999f, 0.15f, 0.499999f, 0.499999f, 0.7f, 0.774999f, 0.925f, 1.f, 1.f, 1.f, 1.f, 1.f,
-					-0.375001f, -0.175001f, -1.49012e-06f, -0.1f, 0.299999f, 0.0499994f, 0.399999f, 0.399999f, 0.6f, 0.674999f, 0.825f, 0.974999f, 0.974999f, 1.f, 1.f, 1.f,
-					-0.475f, -0.274999f, -0.15f, -0.199999f, 0.2f, -0.0499994f, 0.400001f, 0.300001f, 0.500001f, 0.575f, 0.625001f, 0.875001f, 0.875001f, 1.f, 1.f, 1.f,
-					-0.575f, -0.375f, -0.25f, -0.225f, 0.1f, -0.15f, 0.300001f, 0.2f, 0.400001f, 0.475f, 0.525001f, 0.775f, 0.775f, 1.f, 0.95f, 1.f,
-					-0.725f, -0.475f, -0.35f, -0.325f, 0.f, -0.25f, 0.2f, 0.1f, 0.300001f, 0.375f, 0.425f, 0.675f, 0.675f, 0.95f, 0.849999f, 1.f,
-					-0.825f, -0.499999f, -0.45f, -0.425f, -0.1f, -0.35f, 0.1f, 0.f, 0.2f, 0.274999f, 0.325f, 0.575f, 0.575f, 0.849999f, 0.749999f, 1.f,
-					-0.925f, -0.6f, -0.55f, -0.525001f, -0.25f, -0.45f, 0.f, -0.1f, 0.1f, 0.174999f, 0.225f, 0.475f, 0.375f, 0.749999f, 0.649999f, 1.f,
-					-1.f, -0.7f, -0.7f, -0.625001f, -0.35f, -0.55f, -0.1f, -0.2f, 0.f, 0.074999f, 0.125f, 0.375f, 0.274999f, 0.649999f, 0.475f, 0.925f,
-					-1.f, -0.8f, -0.8f, -0.725001f, -0.45f, -0.650001f, -0.2f, -0.300001f, -0.1f, -0.0250012f, 0.0249997f, 0.274999f, 0.174999f, 0.549999f, 0.375f, 0.825f,
-					-1.f, -0.9f, -0.9f, -0.825001f, -0.55f, -0.750001f, -0.300001f, -0.400001f, -0.2f, -0.125001f, -0.0750005f, 0.174999f, 0.074999f, 0.449999f, 0.274999f, 0.725f,
-					-1.f, -0.9f, -0.999999f, -0.849999f, -0.649999f, -0.749999f, -0.399999f, -0.549999f, -0.299999f, -0.199999f, -0.174999f, 0.125f, -0.0249997f, 0.400001f, 0.175001f, 0.625001f,
-					-1.f, -1.f, -1.f, -0.95f, -0.749999f, -0.849999f, -0.499999f, -0.649999f, -0.399999f, -0.299999f, -0.274999f, 0.0249997f, -0.125f, 0.300001f, 0.0750005f, 0.525001f,
-					-1.f, -1.f, -1.f, -1.f, -0.849999f, -0.95f, -0.6f, -0.749999f, -0.499999f, -0.399999f, -0.375f, -0.0750005f, -0.225f, 0.2f, -0.0249997f, 0.425f,
-					-1.f, -1.f, -1.f, -1.f, -0.95f, -1.f, -0.7f, -0.849999f, -0.6f, -0.499999f, -0.475f, -0.175001f, -0.325f, 0.1f, -0.1f, 0.325f,
-					-1.f, -1.f, -1.f, -1.f, -1.f, -1.f, -0.8f, -0.95f, -0.7f, -0.6f, -0.575f, -0.275001f, -0.425f, 0.f, -0.2f, 0.225f;
+			-0.125f, -0.0500008f, 0.199999f, 0.1f, 0.499999f, 0.25f, 0.6f, 0.6f, 0.8f, 0.874999f, 1.f, 1.f, 1.f, 1.f, 1.f, 1.f,
+			-0.275001f, -0.150001f, 0.0999987f, 0.f, 0.399999f, 0.15f, 0.499999f, 0.499999f, 0.7f, 0.774999f, 0.925f, 1.f, 1.f, 1.f, 1.f, 1.f,
+			-0.375001f, -0.175001f, -1.49012e-06f, -0.1f, 0.299999f, 0.0499994f, 0.399999f, 0.399999f, 0.6f, 0.674999f, 0.825f, 0.974999f, 0.974999f, 1.f, 1.f, 1.f,
+			-0.475f, -0.274999f, -0.15f, -0.199999f, 0.2f, -0.0499994f, 0.400001f, 0.300001f, 0.500001f, 0.575f, 0.625001f, 0.875001f, 0.875001f, 1.f, 1.f, 1.f,
+			-0.575f, -0.375f, -0.25f, -0.225f, 0.1f, -0.15f, 0.300001f, 0.2f, 0.400001f, 0.475f, 0.525001f, 0.775f, 0.775f, 1.f, 0.95f, 1.f,
+			-0.725f, -0.475f, -0.35f, -0.325f, 0.f, -0.25f, 0.2f, 0.1f, 0.300001f, 0.375f, 0.425f, 0.675f, 0.675f, 0.95f, 0.849999f, 1.f,
+			-0.825f, -0.499999f, -0.45f, -0.425f, -0.1f, -0.35f, 0.1f, 0.f, 0.2f, 0.274999f, 0.325f, 0.575f, 0.575f, 0.849999f, 0.749999f, 1.f,
+			-0.925f, -0.6f, -0.55f, -0.525001f, -0.25f, -0.45f, 0.f, -0.1f, 0.1f, 0.174999f, 0.225f, 0.475f, 0.375f, 0.749999f, 0.649999f, 1.f,
+			-1.f, -0.7f, -0.7f, -0.625001f, -0.35f, -0.55f, -0.1f, -0.2f, 0.f, 0.074999f, 0.125f, 0.375f, 0.274999f, 0.649999f, 0.475f, 0.925f,
+			-1.f, -0.8f, -0.8f, -0.725001f, -0.45f, -0.650001f, -0.2f, -0.300001f, -0.1f, -0.0250012f, 0.0249997f, 0.274999f, 0.174999f, 0.549999f, 0.375f, 0.825f,
+			-1.f, -0.9f, -0.9f, -0.825001f, -0.55f, -0.750001f, -0.300001f, -0.400001f, -0.2f, -0.125001f, -0.0750005f, 0.174999f, 0.074999f, 0.449999f, 0.274999f, 0.725f,
+			-1.f, -0.9f, -0.999999f, -0.849999f, -0.649999f, -0.749999f, -0.399999f, -0.549999f, -0.299999f, -0.199999f, -0.174999f, 0.125f, -0.0249997f, 0.400001f, 0.175001f, 0.625001f,
+			-1.f, -1.f, -1.f, -0.95f, -0.749999f, -0.849999f, -0.499999f, -0.649999f, -0.399999f, -0.299999f, -0.274999f, 0.0249997f, -0.125f, 0.300001f, 0.0750005f, 0.525001f,
+			-1.f, -1.f, -1.f, -1.f, -0.849999f, -0.95f, -0.6f, -0.749999f, -0.499999f, -0.399999f, -0.375f, -0.0750005f, -0.225f, 0.2f, -0.0249997f, 0.425f,
+			-1.f, -1.f, -1.f, -1.f, -0.95f, -1.f, -0.7f, -0.849999f, -0.6f, -0.499999f, -0.475f, -0.175001f, -0.325f, 0.1f, -0.1f, 0.325f,
+			-1.f, -1.f, -1.f, -1.f, -1.f, -1.f, -0.8f, -0.95f, -0.7f, -0.6f, -0.575f, -0.275001f, -0.425f, 0.f, -0.2f, 0.225f;
 			return expected_tsdf_field01;
 		}();
+
+static eig::MatrixXf expected_tsdf_EWA_voxel_space_inclusive_1 =
+		[] {
+			eig::MatrixXf exp(16,16);
+			exp <<
+			0.141862, 0.00275766, 0.0811198, 0.163639, 0.226404, 0.31303, 0.404172, 0.473615, 0.557036, 0.648832, 0.71873, 0.806699, 0.887528, 0.934625, 0.983267, 0.997929,
+			-0.0399859, -0.062468, 0.0136947, 0.0930131, 0.161298, 0.248675, 0.337132, 0.400155, 0.488666, 0.58037, 0.643468, 0.739822, 0.826796, 0.886504, 0.957229, 0.981247,
+			-0.195677, -0.129732, -0.0519314, 0.015705, 0.0977758, 0.182163, 0.258766, 0.334487, 0.421963, 0.501799, 0.577494, 0.673303, 0.754415, 0.834151, 0.910989, 0.95202,
+			-0.272524, -0.198234, -0.118772, -0.0542549, 0.0304306, 0.118377, 0.180463, 0.267015, 0.355635, 0.42182, 0.511662, 0.606545, 0.672968, 0.766507, 0.853429, 0.910588,
+			-0.339743, -0.26704, -0.194309, -0.11923, -0.0352409, 0.0445354, 0.113788, 0.195748, 0.283548, 0.348348, 0.4417, 0.532979, 0.596774, 0.697306, 0.790259, 0.858437,
+			-0.411366, -0.331683, -0.268946, -0.184914, -0.102941, -0.0326724, 0.0434938, 0.128686, 0.204281, 0.279467, 0.374989, 0.450994, 0.529773, 0.631151, 0.708921, 0.794678,
+			-0.479146, -0.404704, -0.333917, -0.252526, -0.170861, -0.107129, -0.0251701, 0.0611528, 0.124458, 0.209762, 0.305608, 0.370594, 0.460834, 0.560865, 0.627865, 0.727754,
+			-0.547269, -0.480507, -0.401461, -0.324133, -0.245236, -0.175393, -0.0964423, -0.0112473, 0.0517055, 0.142111, 0.232796, 0.292953, 0.388764, 0.485126, 0.552303, 0.657007,
+			-0.615209, -0.549508, -0.469493, -0.390627, -0.325922, -0.244507, -0.165792, -0.0924555, -0.0131655, 0.0703581, 0.150918, 0.227169, 0.32011, 0.402039, 0.485378, 0.5871,
+			-0.691302, -0.615362, -0.53999, -0.461399, -0.393988, -0.31775, -0.235454, -0.170307, -0.0857494, 0.000604492, 0.0682522, 0.156155, 0.24968, 0.31887, 0.412996, 0.515059,
+			-0.761653, -0.684129, -0.609866, -0.539482, -0.46542, -0.388622, -0.308163, -0.24104, -0.157872, -0.0712946, -0.00720363, 0.0838495, 0.178561, 0.240487, 0.340718, 0.435307,
+			-0.827312, -0.757091, -0.682132, -0.613456, -0.53376, -0.460505, -0.389161, -0.312344, -0.230205, -0.149825, -0.0772193, 0.00934458, 0.0950948, 0.169895, 0.268982, 0.352091,
+			-0.891557, -0.828364, -0.753432, -0.683243, -0.60995, -0.531207, -0.467229, -0.383394, -0.299165, -0.231779, -0.149044, -0.0607131, 0.0133653, 0.0997379, 0.196014, 0.265668,
+			-0.950573, -0.892415, -0.82599, -0.754709, -0.683247, -0.604803, -0.535658, -0.456395, -0.371698, -0.310047, -0.222182, -0.13091, -0.0674217, 0.0281761, 0.124028, 0.18436,
+			-0.982555, -0.931911, -0.885617, -0.827715, -0.753718, -0.685699, -0.608594, -0.528099, -0.452323, -0.3813, -0.295989, -0.210528, -0.14167, -0.0470904, 0.0411459, 0.106769,
+			-0.99354, -0.963401, -0.943921, -0.899975, -0.823976, -0.761592, -0.682501, -0.603318, -0.535535, -0.45533, -0.368336, -0.291704, -0.211805, -0.119077, -0.0431902, 0.0356053;
+			return exp;
+		}();
+
+static eig::MatrixXf expected_tsdf_EWA_voxel_space_inclusive_2 =
+		[] {
+			eig::MatrixXf exp(16,16);
+			exp <<
+			1, 1, 1, 1, 1, 1, 0.992266, 0.985577, 0.986085, 0.985577, 0.991991, 1, 1, 1, 1, 1,
+			1, 1, 1, 1, 1, 1, 0.968809, 0.955317, 0.951856, 0.948307, 0.981735, 0.991366, 0.99424, 1, 1, 1,
+			1, 1, 1, 1, 1, 0.989848, 0.926985, 0.909365, 0.881693, 0.885677, 0.950849, 0.961535, 0.985217, 1, 1, 1,
+			1, 1, 1, 1, 1, 0.95679, 0.869823, 0.832531, 0.781519, 0.837487, 0.887433, 0.917018, 0.968319, 0.993517, 0.995193, 0.997331,
+			1, 1, 1, 1, 1, 0.865559, 0.804657, 0.730212, 0.690475, 0.761989, 0.825887, 0.85996, 0.949024, 0.970177, 0.976891, 1,
+			1, 1, 1, 1, 0.933775, 0.791723, 0.694042, 0.615982, 0.637474, 0.684555, 0.750018, 0.818337, 0.879465, 0.931692, 0.954901, 0.999999,
+			1, 1, 1, 1, 0.804177, 0.704671, 0.590452, 0.490458, 0.545537, 0.631603, 0.668391, 0.750986, 0.822039, 0.875941, 0.943308, 0.974051,
+			1, 1, 1, 0.910783, 0.741131, 0.555602, 0.444053, 0.420227, 0.484538, 0.549974, 0.618317, 0.670214, 0.772172, 0.815476, 0.887087, 0.94033,
+			1, 1, 1, 0.761142, 0.604739, 0.434862, 0.29044, 0.337521, 0.431096, 0.468351, 0.543552, 0.622012, 0.674053, 0.765419, 0.812507, 0.876763,
+			1, 1, 0.887813, 0.711564, 0.458675, 0.267091, 0.21301, 0.284518, 0.331735, 0.418295, 0.470195, 0.550819, 0.615429, 0.665506, 0.759493, 0.803806,
+			1, 1, 0.765222, 0.504862, 0.347478, 0.0904254, 0.137511, 0.211005, 0.268315, 0.318391, 0.421987, 0.474007, 0.538078, 0.612493, 0.659579, 0.750937,
+			1, 0.864866, 0.650158, 0.370039, 0.0901499, -0.00949123, 0.0844951, 0.131693, 0.189924, 0.270173, 0.322086, 0.415379, 0.465479, 0.536985, 0.609508, 0.651116,
+			1, 0.755284, 0.405037, 0.209014, -0.126834, -0.0625046, -0.0104733, 0.0682717, 0.118365, 0.197172, 0.273955, 0.315558, 0.412473, 0.45955, 0.53055, 0.60295,
+			0.841941, 0.5888, 0.264144, -0.0867616, -0.209506, -0.132757, -0.068344, -0.0143198, 0.0701576, 0.122062, 0.196177, 0.265459, 0.312533, 0.409457, 0.451063, 0.527836,
+			0.829985, 0.30527, 0.0706007, -0.278941, -0.262515, -0.215434, -0.149003, -0.0816542, -0.0106163, 0.073908, 0.115513, 0.195225, 0.259527, 0.310434, 0.40292, 0.45483,
+			0.527488, 0.254249, -0.263649, -0.409521, -0.338017, -0.268381, -0.231591, -0.14063, -0.0779629, -0.023906, 0.0654381, 0.112519, 0.190265, 0.25101, 0.317795, 0.406611;
+			return exp;
+		}();
+
 } //namespace test_data
