+++ tests/test_tsdf.cpp
@@ -35,9 +35,6 @@
 #include "data/test_data_tsdf.hpp"
 
 //test targets
-#include "../src/tsdf/tsdf.hpp"
-#include "../src/tsdf/ewa.hpp"
-
 #include "../src/tsdf/generator.hpp"
 #include "../src/math/typedefs.hpp"
 #include "../src/math/almost_equal.hpp"
@@ -46,7 +43,58 @@
 
 namespace eig = Eigen;
 
-BOOST_AUTO_TEST_CASE(test_EWA_2D_generation01) {
+
+BOOST_AUTO_TEST_CASE(test_tsdf_generation_no_interpolation_1) {
+	math::MatrixXus depth_image;
+	bool image_read = read_image_helper(depth_image, "zigzag2_depth_00108.png");
+	BOOST_REQUIRE(image_read);
+	eig::Matrix3f camera_intrinsic_matrix;
+	camera_intrinsic_matrix <<
+			700.0f, 0.0f, 320.0f,
+			0.0f, 700.0f, 240.0f,
+			0.0f, 0.0f, 1.0f;
+	tsdf::Parameters2d parameters(
+				0.001f, //depth unit ratio
+				camera_intrinsic_matrix, //projection matrix
+				0.05f, //near clipping distance
+				math::Vector2i(-8, 144), //offset
+				math::Vector2i(16, 16), //shape
+				0.004f, //voxel size
+				20, //narrow band width
+				tsdf::InterpolationMethod::NONE
+				);
+	tsdf::Generator2d generator(parameters);
+	eig::MatrixXf field = generator.generate(depth_image, eig::Matrix4f::Identity(), 200);
+	BOOST_REQUIRE(math::almost_equal_verbose(field, test_data::expected_tsdf_field01, 1e-6f));
+}
+
+BOOST_AUTO_TEST_CASE(test_tsdf_generation_no_interpolation_2) {
+	math::MatrixXus depth_image;
+	bool image_read = read_image_helper(depth_image, "zigzag2_depth_00108.png");
+	BOOST_REQUIRE(image_read);
+	eig::Matrix3f camera_intrinsic_matrix;
+	camera_intrinsic_matrix <<
+			700.0f, 0.0f, 320.0f,
+			0.0f, 700.0f, 240.0f,
+			0.0f, 0.0f, 1.0f;
+	eig::Matrix4f camera_extrinsic_matrix = eig::Matrix4f::Identity();
+	camera_extrinsic_matrix(2, 3) = 0.004;
+	tsdf::Parameters2d parameters(
+					0.001f, //depth unit ratio
+					camera_intrinsic_matrix, //projection matrix
+					0.05f, //near clipping distance
+					math::Vector2i(-8, 144), //offset
+					math::Vector2i(16, 16), //shape
+					0.004f, //voxel size
+					20, //narrow band width
+					tsdf::InterpolationMethod::NONE
+					);
+	tsdf::Generator2d generator(parameters);
+	eig::MatrixXf field = generator.generate(depth_image, camera_extrinsic_matrix, 200);
+	BOOST_REQUIRE(math::almost_equal_verbose(field, test_data::expected_tsdf_field02, 1e-6f));
+}
+
+BOOST_AUTO_TEST_CASE(test_tsdf_generation_EWA_2D_image_space_1) {
 
 	math::MatrixXus depth_image = math::MatrixXus::Constant(3, 640, USHRT_MAX);
 	depth_image.block(0, 399, 3, 18) = test_data::depth_image_region;
@@ -57,8 +105,6 @@ BOOST_AUTO_TEST_CASE(test_EWA_2D_generation01) {
 			0.0f, 700.0f, 240.0f,
 			0.0f, 0.0f, 1.0f;
 
-	eig::Vector2i field_shape;
-	field_shape << 16, 16;
 	tsdf::Parameters2d parameters(
 			0.001f, //depth unit ratio
 			camera_intrinsic_matrix, //projection matrix
@@ -76,7 +122,7 @@ BOOST_AUTO_TEST_CASE(test_EWA_2D_generation01) {
 	BOOST_REQUIRE(math::almost_equal_verbose(field, test_data::out_sdf_field, 1e-6f));
 }
 
-BOOST_AUTO_TEST_CASE(test_EWA_2D_generation02) {
+BOOST_AUTO_TEST_CASE(test_tsdf_generation_EWA_2D_image_space_2) {
 	math::MatrixXus depth_image;
 	bool image_read = read_image_helper(depth_image, "zigzag2_depth_00108.png");
 	BOOST_REQUIRE(image_read);
@@ -112,23 +158,23 @@ BOOST_AUTO_TEST_CASE(test_EWA_2D_generation02) {
 		field_chunk = field.block(chunk_y_start, chunk_x_start, chunk_size, chunk_size);
 	} else {
 		tsdf::Parameters2d parameters(
-						0.001f, //depth unit ratio
-						camera_intrinsic_matrix, //projection matrix
-						0.05f, //near clipping distance
-						offset_chunk,
-						math::Vector2i(chunk_size, chunk_size), //shape
-						0.004f, //voxel size
-						20, //narrow band width
-						tsdf::InterpolationMethod::EWA_IMAGE_SPACE,
-						1.0f // gaussian covariance scale
-						);
+				0.001f, //depth unit ratio
+				camera_intrinsic_matrix, //projection matrix
+				0.05f, //near clipping distance
+				offset_chunk,
+				math::Vector2i(chunk_size, chunk_size), //shape
+				0.004f, //voxel size
+				20, //narrow band width
+				tsdf::InterpolationMethod::EWA_IMAGE_SPACE,
+				1.0f // gaussian covariance scale
+				);
 		tsdf::Generator2d generator(parameters);
 		field_chunk = generator.generate(depth_image, eig::Matrix4f::Identity(), 200);
 	}
 	BOOST_REQUIRE(math::almost_equal_verbose(field_chunk, test_data::out_sdf_chunk, 1e-6f));
 }
 
-BOOST_AUTO_TEST_CASE(test_EWA_2D_generation05) {
+BOOST_AUTO_TEST_CASE(test_tsdf_generation_EWA_2D_voxel_space_inclusive_1) {
 	math::MatrixXus depth_image;
 	bool image_read = read_image_helper(depth_image, "zigzag2_depth_00108.png");
 	BOOST_REQUIRE(image_read);
@@ -139,49 +185,48 @@ BOOST_AUTO_TEST_CASE(test_EWA_2D_generation05) {
 			700.0f, 0.0f, 320.0f,
 			0.0f, 700.0f, 240.0f,
 			0.0f, 0.0f, 1.0f;
-	eig::Vector3i offset;
-	offset << -256, -256, 0;
+	math::Vector2i offset(-256, 0);
 	int chunk_x_start = 210, chunk_y_start = 103;
 	int chunk_size = 16;
-	eig::Vector3i offset_chunk_from_image;
-	offset_chunk_from_image << chunk_x_start, 0.0f, chunk_y_start;
-	eig::Vector3i offset_chunk = offset + offset_chunk_from_image;
+	math::Vector2i offset_chunk_from_image(chunk_x_start, chunk_y_start);
+	math::Vector2i offset_chunk = offset + offset_chunk_from_image;
 
 	eig::MatrixXf field_chunk;
 	if (test_full_image) {
-		eig::MatrixXf field = tsdf::generate_TSDF_2D_EWA_TSDF_inclusive(
-				200, // y coord
-				depth_image,
+		tsdf::Parameters2d parameters(
 				0.001f, //depth unit ratio
-				camera_intrinsic_matrix,
-				eig::Matrix4f::Identity(), //camera pose
-				offset,
-				512, //field size
+				camera_intrinsic_matrix, //projection matrix
+				0.05f, //near clipping distance
+				offset_chunk,
+				math::Vector2i(512, 512), //shape
 				0.004f, //voxel size
-				20 // narrow band width
+				20, //narrow band width
+				tsdf::InterpolationMethod::EWA_VOXEL_SPACE_INCLUSIVE,
+				1.0f // gaussian covariance scale
 				);
-
+		tsdf::Generator2d generator(parameters);
+		eig::MatrixXf field = generator.generate(depth_image, eig::Matrix4f::Identity(), 200);
 		field_chunk = field.block(chunk_y_start, chunk_x_start, chunk_size, chunk_size);
 	} else {
-		field_chunk = tsdf::generate_TSDF_2D_EWA_TSDF_inclusive(
-				200, // y coord
-				depth_image,
+		tsdf::Parameters2d parameters(
 				0.001f, //depth unit ratio
-				camera_intrinsic_matrix,
-				eig::Matrix4f::Identity(), //camera pose
+				camera_intrinsic_matrix, //projection matrix
+				0.05f, //near clipping distance
 				offset_chunk,
-				chunk_size, //field size
+				math::Vector2i(chunk_size, chunk_size), //shape
 				0.004f, //voxel size
-				20 // narrow band width
+				20, //narrow band width
+				tsdf::InterpolationMethod::EWA_VOXEL_SPACE_INCLUSIVE,
+				1.0f // gaussian covariance scale
 				);
+		tsdf::Generator2d generator(parameters);
+		field_chunk = generator.generate(depth_image, eig::Matrix4f::Identity(), 200);
 	}
 
-	console::print_initializer_list(std::cout,field_chunk,true);
-	//TODO: add data for this test
-	//BOOST_REQUIRE(math::almost_equal_verbose(field_chunk, test_data::out_sdf_chunk, 1e-6f));
+	BOOST_REQUIRE(math::almost_equal_verbose(field_chunk, test_data::expected_tsdf_EWA_voxel_space_inclusive_1, 1e-6f));
 }
 
-BOOST_AUTO_TEST_CASE(test_EWA_2D_generation06) {
+BOOST_AUTO_TEST_CASE(test_tsdf_generation_EWA_2D_voxel_space_inclusive_2) {
 	math::MatrixXus depth_image;
 	bool image_read = read_image_helper(depth_image, "zigzag_depth_00064.png");
 	BOOST_REQUIRE(image_read);
@@ -192,162 +237,98 @@ BOOST_AUTO_TEST_CASE(test_EWA_2D_generation06) {
 			700.0f, 0.0f, 320.0f,
 			0.0f, 700.0f, 240.0f,
 			0.0f, 0.0f, 1.0f;
-	eig::Vector3i offset;
-	offset << -256, -256, 480;
+	math::Vector2i offset(-256, 480);
 	int chunk_x_start = 24, chunk_y_start = 10;
 	int chunk_size = 16;
-	eig::Vector3i offset_chunk_from_image;
-	offset_chunk_from_image << chunk_x_start, 0.0f, chunk_y_start;
-	eig::Vector3i offset_chunk = offset + offset_chunk_from_image;
+	math::Vector2i offset_chunk_from_image(chunk_x_start, chunk_y_start);
+	math::Vector2i offset_chunk = offset + offset_chunk_from_image;
 
 	eig::MatrixXf field_chunk;
 	if (test_full_image) {
-
-		eig::MatrixXf field = tsdf::generate_TSDF_2D_EWA_TSDF_inclusive(
-				200, // y coord
-				depth_image,
-				0.001f, //depth unit ratio
-				camera_intrinsic_matrix,
-				eig::Matrix4f::Identity(), //camera pose
-				offset,
-				512, //field size
+		tsdf::Parameters2d parameters(0.001f, //depth unit ratio
+				camera_intrinsic_matrix, //projection matrix
+				0.05f, //near clipping distance
+				offset_chunk, //offset of scene from world origin
+				math::Vector2i(512, 512), //shape
 				0.004f, //voxel size
-				20 // narrow band width
+				20, //narrow band width
+				tsdf::InterpolationMethod::EWA_VOXEL_SPACE_INCLUSIVE,
+				1.0f // gaussian covariance scale
 				);
+		tsdf::Generator2d generator(parameters);
+		eig::MatrixXf field = generator.generate(depth_image, eig::Matrix4f::Identity(), 200);
 		field_chunk = field.block(chunk_y_start, chunk_x_start, chunk_size, chunk_size);
 	} else {
-		field_chunk = tsdf::generate_TSDF_2D_EWA_TSDF_inclusive(
-				200, // y coord
-				depth_image,
+		tsdf::Parameters2d parameters(
 				0.001f, //depth unit ratio
-				camera_intrinsic_matrix,
-				eig::Matrix4f::Identity(), //camera pose
+				camera_intrinsic_matrix, //projection matrix
+				0.05f, //near clipping distance
 				offset_chunk,
-				chunk_size, //field size
+				math::Vector2i(chunk_size, chunk_size), //shape
 				0.004f, //voxel size
-				20 // narrow band width
+				20, //narrow band width
+				tsdf::InterpolationMethod::EWA_VOXEL_SPACE_INCLUSIVE,
+				1.0f // gaussian covariance scale
 				);
+		tsdf::Generator2d generator(parameters);
+		field_chunk = generator.generate(depth_image, eig::Matrix4f::Identity(), 200);
 	}
-	//TODO test against data
-	//BOOST_REQUIRE(math::almost_equal_verbose(field_chunk, test_data::out_sdf_chunk, 1e-6f));
+	BOOST_REQUIRE(math::almost_equal_verbose(field_chunk, test_data::expected_tsdf_EWA_voxel_space_inclusive_2, 1e-6f));
 }
 
-BOOST_AUTO_TEST_CASE(test_EWA_3D_generation01) {
+BOOST_AUTO_TEST_CASE(test_tsdf_generation_EWA_3D_image_space_1) {
 	math::MatrixXus depth_image;
 	bool image_read = read_image_helper(depth_image, "zigzag2_depth_00108.png");
 	BOOST_REQUIRE(image_read);
 
-	eig::Vector3i offset;
-	offset << -46, -8, 105; //zigzag2-108
-
-	eig::Vector3i field_size;
-	field_size << 16, 1, 16; //zigzag2-108
-
+	math::Vector3i offset(-46, -8, 105);
+	math::Vector3i field_shape(16, 1, 16);
 	eig::Matrix3f camera_intrinsic_matrix;
 	camera_intrinsic_matrix <<
 			700.0f, 0.0f, 320.0f,
 			0.0f, 700.0f, 240.0f,
 			0.0f, 0.0f, 1.0f;
 
-	eig::Tensor<float, 3> field = tsdf::generate_TSDF_3D_EWA_image(
-			depth_image,
-			0.001f, //depth unit ratio
-			camera_intrinsic_matrix,
-			eig::Matrix4f::Identity(), //camera pose
-			offset,
-			field_size, //field size
-			0.004f, //voxel size
-			20 // narrow band width
-			);
+	tsdf::Parameters3d parameters(0.001f, //depth unit ratio
+						camera_intrinsic_matrix, //projection matrix
+						0.05f, //near clipping distance
+						offset, //offset of scene from world origin
+						field_shape, //dimensions of the voxel grid
+						0.004f, //voxel size
+						20, //narrow band width
+						tsdf::InterpolationMethod::EWA_IMAGE_SPACE,
+						1.0f // gaussian covariance scale
+						);
+	tsdf::Generator3d generator(parameters);
+	eig::Tensor<float, 3> field = generator.generate(depth_image, eig::Matrix4f::Identity());
 
 	BOOST_REQUIRE(math::almost_equal_verbose(field, test_data::TSDF_slice01, 1e-6f));
 }
 
-BOOST_AUTO_TEST_CASE(test_EWA_3D_generation02) {
+BOOST_AUTO_TEST_CASE(test_tsdf_generation_EWA_3D_image_space_2) {
 	math::MatrixXus depth_image;
 	bool image_read = read_image_helper(depth_image, "zigzag2_depth_00108.png");
 	BOOST_REQUIRE(image_read);
 
-	eig::Vector3i offset;
-	offset << -46, -8, 105; //zigzag2-108
-
-	eig::Vector3i field_size;
-	field_size << 16, 1, 16; //zigzag2-108
-
+	math::Vector3i offset(-46, -8, 105);
+	math::Vector3i field_shape(16, 1, 16); //zigzag2-108
 	eig::Matrix3f camera_intrinsic_matrix;
 	camera_intrinsic_matrix <<
 			700.0f, 0.0f, 320.0f,
 			0.0f, 700.0f, 240.0f,
 			0.0f, 0.0f, 1.0f;
-
-	eig::Tensor<float, 3> field = tsdf::generate_TSDF_3D_EWA_image(
-			depth_image,
-			0.001f, //depth unit ratio
-			camera_intrinsic_matrix,
-			eig::Matrix4f::Identity(), //camera pose
-			offset,
-			field_size, //field size
-			0.004f, //voxel size
-			20, // narrow band width
-			0.5f
-			);
+	tsdf::Parameters3d parameters(0.001f, //depth unit ratio
+							camera_intrinsic_matrix, //projection matrix
+							0.05f, //near clipping distance
+							offset, //offset of scene from world origin
+							field_shape, //dimensions of the voxel grid
+							0.004f, //voxel size
+							20, //narrow band width
+							tsdf::InterpolationMethod::EWA_IMAGE_SPACE,
+							0.5f // gaussian covariance scale
+							);
+	tsdf::Generator3d generator(parameters);
+	eig::Tensor<float, 3> field = generator.generate(depth_image, eig::Matrix4f::Identity());
 	BOOST_REQUIRE(math::almost_equal_verbose(field, test_data::TSDF_slice02, 1e-5f));
 }
 
-// Following tests are for regular tsdf generation without interpolation
-BOOST_AUTO_TEST_CASE(test_TSDF_2D_generation01) {
-	math::MatrixXus depth_image;
-	bool image_read = read_image_helper(depth_image, "zigzag2_depth_00108.png");
-	BOOST_REQUIRE(image_read);
-	eig::Matrix3f camera_intrinsic_matrix;
-	camera_intrinsic_matrix <<
-			700.0f, 0.0f, 320.0f,
-			0.0f, 700.0f, 240.0f,
-			0.0f, 0.0f, 1.0f;
-	eig::Vector3i offset;
-	offset << -8, -8, 144;
-	int field_size = 16;
-	eig::MatrixXf field = tsdf::generate_TSDF_2D(
-			200, // y coord
-			depth_image,
-			0.001f, //depth unit ratio
-			camera_intrinsic_matrix,
-			eig::Matrix4f::Identity(), //camera pose
-			offset,
-			field_size, //field size
-			0.004f, //voxel size
-			20, // narrow band width
-			-999.f
-			);
-	BOOST_REQUIRE(math::almost_equal_verbose(field, test_data::expected_tsdf_field01, 1e-6f));
-}
-
-BOOST_AUTO_TEST_CASE(test_TSDF_2D_generation02) {
-	math::MatrixXus depth_image;
-	bool image_read = read_image_helper(depth_image, "zigzag2_depth_00108.png");
-	BOOST_REQUIRE(image_read);
-	eig::Matrix3f camera_intrinsic_matrix;
-	camera_intrinsic_matrix <<
-			700.0f, 0.0f, 320.0f,
-			0.0f, 700.0f, 240.0f,
-			0.0f, 0.0f, 1.0f;
-	eig::Matrix4f camera_extrinsic_matrix = eig::Matrix4f::Identity();
-	camera_extrinsic_matrix(2, 3) = 0.004;
-
-	eig::Vector3i offset;
-	offset << -8, -8, 144;
-	int field_size = 16;
-	eig::MatrixXf field = tsdf::generate_TSDF_2D(
-			200, // y coord
-			depth_image,
-			0.001f, //depth unit ratio
-			camera_intrinsic_matrix,
-			camera_extrinsic_matrix, //camera pose
-			offset,
-			field_size, //field size
-			0.004f, //voxel size
-			20, // narrow band width
-			-999.f
-	);
-	BOOST_REQUIRE(math::almost_equal_verbose(field, test_data::expected_tsdf_field02, 1e-6f));
-}
